API Refactor: carto.widget
