+++ src/lib/widget/index.ts
@@ -1,0 +1,2 @@
+export { CategoryWidget } from './category/CategoryWidget';
+export { FormulaWidget } from './formula/FormulaWidget';
