--- src/lib/widgets/index.ts
@@ -1,2 +1,0 @@
-export { Category } from './category/Category';
-export { Formula } from './formula/Formula';
